Conflicts:
	ProjectFlightSchool/Debug/mapNodeLog.txt
	ProjectFlightSchool/ProjectFlightSchool/PlayState.cpp
	ProjectFlightSchool/ProjectFlightSchool/ProjectFlightSchool.vcxproj.filters
	ProjectFlightSchool/ProjectFlightSchool/RenderManager.cpp
	ProjectFlightSchool/ProjectFlightSchool/mapNodeLog.txt

+++ ProjectFlightSchool/ProjectFlightSchool/Structs.h
@@ -65,7 +65,7 @@ struct EvSyncEnemy
 struct EvMeleeHit
 {
 	unsigned int			ID;
-	unsigned int			damage;
+	float					damage;
 	float					knockBack;
 	DirectX::XMFLOAT3		direction;
 };
